Remove web-encoding, it is part of global jest config
