+++ plugins/spreadsheet-view/src/SpreadsheetView/importAdapters/STARFusionImport.test.ts
@@ -1,9 +1,9 @@
 import { promises as fsPromises } from 'fs'
 import path from 'path'
+
+// locals
 import { parseSTARFusionBuffer } from './STARFusionImport'
-import { TextDecoder } from 'web-encoding'
 import SpreadsheetModel from '../models/Spreadsheet'
-window.TextDecoder = TextDecoder
 
 test('starfusion import', async () => {
   const filepath = path.join(
